Update Az-MX.py
This commit adds the building of the second tunnel configuration in the Meraki dashboard. Although, this script still needs to have the Azure sitedata also account for sites with a single ISP. (Currently script assumes sites have dual uplinks). And the script needs to add the backup tag and not the primary vWAN tag for the backupazure tunnel to instance1.